+++ NEWS.md
@@ -1,6 +1,7 @@
 # caracas 2.1.1
 
 * `print(..., method = "compactascii")`: column vector as transposed row vector
+* `LUdecomposition()` added.
 
 # caracas 2.1.0